+++ apps/test/__tests__/main_file_output.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from '@jest/globals';
 const fs = require('fs');
 
-it('boostest main file output correctly', () => {
+it.only('boostest main file output correctly', () => {
   try {
     const data = fs.readFileSync('./src/main_test_test_data.ts', 'utf-8');
