+++ crates/boostest/src/boostest_utils.rs
@@ -1,2 +1,5 @@
 pub mod ast_utils;
+pub mod module_resolver;
+pub mod setting;
+pub mod task;
 pub mod utils;
